+++ modules/AWSPowerShell/Cmdlets/RDS/AWS.Tools.RDS.Completers.psm1
@@ -343,7 +343,8 @@ $RDS_SelectMap = @{
                "Stop-RDSActivityStream",
                "Stop-RDSDBCluster",
                "Stop-RDSDBInstance",
-               "Stop-RDSDBInstanceAutomatedBackupsReplication")
+               "Stop-RDSDBInstanceAutomatedBackupsReplication",
+               "Convert-RDSReadReplicaToNewPrimary")
 }
 
 _awsArgumentCompleterRegistration $RDS_SelectCompleters $RDS_SelectMap