+++ modules/AWSPowerShell/Cmdlets/SSOAdmin/Basic/New-SSOADMNInstanceAccessControlAttributeConfiguration-Cmdlet.cs
@@ -32,6 +32,11 @@ namespace Amazon.PowerShell.Cmdlets.SSOADMN
     /// Web Services SSO instance. You can also specify new attributes to add to your ABAC
     /// configuration during the enabling process. For more information about ABAC, see <a href="/singlesignon/latest/userguide/abac.html">Attribute-Based Access Control</a>
     /// in the <i>Amazon Web Services SSO User Guide</i>.
+    /// 
+    ///  <note><para>
+    /// After a successful response, call <code>DescribeInstanceAccessControlAttributeConfiguration</code>
+    /// to validate that <code>InstanceAccessControlAttributeConfiguration</code> was created.
+    /// </para></note>
     /// </summary>
     [Cmdlet("New", "SSOADMNInstanceAccessControlAttributeConfiguration", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("None")]